Turn `getProcessGroup` into a method, make everything use this fn
Now this function is public, and a method of `api.Machine`

There were a couple of places that still expected either "process_group" or "fly_process_group", so this normalizes everything (but should retain backwards compatibility!)

This should be tested just to be sure that nothing depends on a specific key in the backend, though :)
