+++ internal/command/machine/status.go
@@ -9,7 +9,8 @@ import (
 
 	"github.com/alecthomas/chroma/quick"
 	"github.com/spf13/cobra"
-	"github.com/superfly/flyctl/api"
+	"github.com/superfly/flyctl/flaps"
+	"github.com/superfly/flyctl/internal/appconfig"
 	"github.com/superfly/flyctl/internal/command"
 	"github.com/superfly/flyctl/internal/flag"
 	"github.com/superfly/flyctl/internal/render"
@@ -45,25 +46,6 @@ func newStatus() *cobra.Command {
 	return cmd
 }
 
-func getProcessGroup(m *api.Machine) string {
-	// backwards compatible process_group getter.
-	// from poking around, "fly_process_group" used to be called "process_group"
-	// and since it's a metadata value, it's like a screenshot.
-	// so we have 3 scenarios
-	// - machines with only 'process_group'
-	// - machines with both 'process_group' and 'fly_process_group'
-	// - machines with only 'fly_process_group'
-
-	processGroup := m.Config.Metadata["process_group"]
-	flyProcessGroup := m.Config.Metadata[api.MachineConfigMetadataKeyFlyProcessGroup]
-
-	if flyProcessGroup != "" {
-		return flyProcessGroup
-	}
-
-	return processGroup
-}
-
 func runMachineStatus(ctx context.Context) (err error) {
 	io := iostreams.FromContext(ctx)
 
@@ -86,7 +68,7 @@ func runMachineStatus(ctx context.Context) (err error) {
 			machine.Name,
 			machine.PrivateIP,
 			machine.Region,
-			getProcessGroup(machine),
+			machine.ProcessGroup(),
 			fmt.Sprint(machine.Config.Guest.CPUKind),
 			fmt.Sprint(machine.Config.Guest.CPUs),
 			fmt.Sprint(machine.Config.Guest.MemoryMB),
